Move zmq-related into messenger dir
zsocket does not belong in top-level

rename zsocket to _socket

move messenger into zeromq dir
